Update postinstall script
