+++ cli/npm/turbo-install/bin/turbo
@@ -1,11 +1,9 @@
 #!/usr/bin/env node
-throw new Error(`turbo: Failed to install correctly
-Make sure you don't have "ignore-scripts" set to true. You can check this with
-"npm config get ignore-scripts". If that returns true you can reset it back to
-false using "npm config set ignore-scripts false" and then reinstall esbuild.
-If you're using npm v7, make sure your package-lock.json file contains either
-"lockfileVersion": 1 or the code "hasInstallScript": true. If it doesn't have
-either of those, then it is likely the case that a known bug in npm v7 has
-corrupted your package-lock.json file. Regenerating your package-lock.json file
-should fix this issue.
-`)
+
+const { generateBinPath } = require("../node-platform");
+
+require("child_process").execFileSync(
+  generateBinPath(),
+  process.argv.slice(2),
+  { stdio: "inherit" }
+);
