+++ cli/npm/turbo-install/package.json
@@ -13,8 +13,22 @@
   },
   "files": [
     "bin",
-    "install.js",
-    "login.js"
+    "node-platform.js",
+    "install.js"
   ],
+  "optionalDependencies": {
+    "turbo-darwin-64": "1.0.20-canary.1",
+    "turbo-darwin-arm64": "1.0.20-canary.1",
+    "turbo-freebsd-64": "1.0.20-canary.1",
+    "turbo-freebsd-arm64": "1.0.20-canary.1",
+    "turbo-linux-32": "1.0.20-canary.1",
+    "turbo-linux-64": "1.0.20-canary.1",
+    "turbo-linux-arm": "1.0.20-canary.1",
+    "turbo-linux-arm64": "1.0.20-canary.1",
+    "turbo-linux-mips64le": "1.0.20-canary.1",
+    "turbo-linux-ppc64le": "1.0.20-canary.1",
+    "turbo-windows-32": "1.0.20-canary.1",
+    "turbo-windows-64": "1.0.20-canary.1"
+  },
   "license": "MPL-2.0"
 }
